Add working directory option
New command-line utility option:

   -d, --working-dir=WORKING-DIR

If specified, use the given directory as working directory.

Follows same syntax as Composer (quick port from there).

There is a clash with the process command -d alias. It's behavior
changes:

Old usage of the -d flag will throw an exception as it now requires an
argument.

Instead use the -n switch which is also in the commons for --dry-run (e.g.
see git).