Add rector.yaml to .dockerignore
This is the rector configuration file for developing on rector itself, and should not be used in runtime.

Without this change, the docker image (which runs with `/rector` as its working directory) would use this `rector.yaml` as the default configuration file. This leads to unexpected effects, such as rector ignoring any `tests/`-directory by default.