Regression test class-name validity
In PseudoNamespaceToNamespaceRectorTest

For class_exists() which triggers the auto-loader, extend the test-case
to keep track of all auto-loaded classes and assert that all class-names
used in auto-loading are valid class-names.

NOTE: The class-names in the PseudoNamespaceToNamespaceRector are not
      meant to be actually existing class-names, but just some that
      might be and there-fore always valid ones.

The problem with invalid class-names is specifically, that they never
exist and trigger the auto-loader always. If they contain two consecutive
backslashes (name-space-separator) then the mapping to the file-system
might merge them to an already included file which will make rector exit
with a fatal error as a class is getting re-defined by the auto-load
include:

> Fatal error: Cannot declare class <class-name>, because the name is
> already in use in ...

Therefore the PseudoNamespaceToNamespaceRector must not trigger
auto-loading on invalid class-names.

Fixtures known to fail:

- fixture4.php.inc
- fixture5.php.inc
- fixture6.php.inc
- var_doc.php.inc

Notes on Phpstan supressions (others should be self-explanatory):

- getAutoloadClassNameAssertion() has no return-type hint by intention
  as static code-analyzers (Phpstorm) won't be able to infer the concrete
  type of the anonymous class returned.

- bogus error for anonymous classes:

        '#Class "AnonymousClassb3d71a09d1e123ac8b2eb26b93f1b476" should have suffix "TestCase" by parent class/interface#'