Fix invalid class-name concatenating w/ name-space
Previously ObjectTypeSpecifier::matchSameNamespacedObjectType() was prone
to create invalid class-names with two consecutive backslashes.

These can lead to fatal errors as class_exists() always returns false
for them while triggering an auto-load which may resolve to a file with
a class-definition that is already defined (as the class-name with a
single backslash would be valid and already loaded).

Rector then exists with a fatal error.

> Fatal error: Cannot declare class <class-name>, because the name is
> already in use in ...

Fix is to prevent this fatal error cause by auto-loading by preventing
consecutive backslashes when concatenating the namespace-object class-name
string.

Either the class is already loaded then class_exists() does not trigger the
auto-loading any longer -or- it does trigger the auto-loading and it can
eventually be loaded and would never clash.

Ref-Issue: #5024