Let test fail again
I realized that I did not provide the test case in #5564 correctly as I was running it in my project.

By changing the existing test it fails again.

Why? Because `OtherWildcardSubscriber` does match the wildcard `*WildcardSubscriber`. That one does not implement an interface.

Because `$this->classManipulator->hasParentMethodOrInterface` just tries to find the first matching class that matches the wildcard, it checks that one always.

That means that the implementation in #5622 is not correct (and naive).