Declare dependency into intl extension
Upon installing rector, I attempted to run the command specified in the
readme file: `vendor/bin/rector init`. However, this command ended up in
a fatal error that was caused by my PHP environment not having the
intl extension installed.

This change adds this dependency to the composer.json file. This would cause
install processes to complain that the system where it's being installed don't
meet all the requirements, thus, avoiding users to encounter the same
fatal error I encountered, providing the composer command it's not run with the
`ignore-platform-reqs`.