Update how_it_works.md
Hey Rector Team. I think I spotted what seems to be a typo at the end in the similar projects section specifically for hhast it says mirations rather than migrations the "g" is missing I think. But I might be wrong. Hopefully it helps. Have a great day. And thanks for this awesome tool. I am learning to use it to it's full power. Thanks again and take care.