[Downgrade 7.0] Strip unnecessary parentheses around expressions (#1617)
The Uniform variable syntax RFC implemented in PHP 7.0 allows wrapping
arbitrary expressions in parentheses, which is not supported by PHP 5.
Since PHP 5 cannot parse that we need to strip them.

And because parenthesization is not part of the AST, we need to force
the code to be re-generated. Rector will re-generate code for AST nodes
that changed. Or we can remove the original node reference to the same effect.
The code generator will only put parentheses where strictly necessary,
like to disambiguate method call from calling a callable property
in `($foo->bar)()`. That should be handled by a different rule.